requirements.txt
kokoro-onnx>=0.4.7
fastrtc[vad]>=0.0.19
groq>=0.22.0
numpy>=2.1.3
langgraph>=0.1.18
langchain-core>=0.1.29
langchain-groq>=0.1.5
loguru>=0.7.3
python-dotenv
soundfile
langchain-tavily
cartesia==3.0.0b2
gradio==5.29.1 
gradio-client==1.10.1
langchain-community
langchain-openai
langchain-cerebras
yfinance
streamlit
streamlit-option-menu
langchain-huggingface
pypdf
sentence-transformers
langchain-chroma
SpeechRecognition
pyowm
firecrawl-py
gTTS
fastapi
fastapi
uvicorn
pydantic-settings
httpx
httpx
pydantic
pydantic-settings
uvicorn
streamlit-option-menu
